pyproject.toml
[project]
name = "a3presentation"
version = "0.1.0"
description = "Template-based PowerPoint presentation generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0,<1.0.0",
    "pydantic>=2.11.0,<3.0.0",
    "pypdf>=6.0.0,<7.0.0",
    "python-docx>=1.2.0,<2.0.0",
    "python-multipart>=0.0.9,<1.0.0",
    "python-pptx>=1.0.2,<2.0.0",
    "uvicorn[standard]>=0.35.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
